refactor(root): move executable files to bin folder
