+++ src/bin/findPackages.js
@@ -1,12 +1,13 @@
+#!/usr/bin/env node
 // @flow
 
 import path from 'path';
 
 import commandLineArgs from 'command-line-args';
 
-import d3DirTree from '../packages/utils/lib/d3DirTree';
+import d3DirTree from '../../packages/utils/lib/d3DirTree';
 // eslint-disable-next-line max-len
-import type { d3DirTreeType } from '../packages/utils/src/definitions/d3DirTree.js.flow';
+import type { d3DirTreeType } from '../../packages/utils/src/definitions/d3DirTree.js.flow';
 
 const { ignore, showInfo } = commandLineArgs([
   {
@@ -24,7 +25,9 @@ const { ignore, showInfo } = commandLineArgs([
   },
 ]);
 
-const packages = d3DirTree(path.resolve(__dirname, './../packages'))
+const packages = d3DirTree(path.resolve(__dirname, '../../packages'), {
+  exclude: ignore.map((ignoreText: string): RegExp => new RegExp(ignoreText)),
+})
   .children.filter(
     ({ data }: d3DirTreeType): boolean => !ignore.includes(data.name),
   )
